refactor(desktop): split preload into per-window-kind bundles (W4)
W4 from the deferred-grouping research doc — a Tier 2 "scaffolding"
change that introduces a per-surface preload seam without re-shaping
user-visible behavior. Surface scope is unchanged: every kind exposes
the identical 15-key LightfastBridge today; the value is the structural
slot for future trust-boundary work (HUD lockdown, in-app web-view host,
plugin host) without paying the build re-shape cost at that point. The
research doc explicitly notes the merge cost of doing this later is
identical, so we claim the slot now.

Authoring layer:
- New apps/desktop/src/preload/build-bridge.ts owns bridge construction
  and exposure via a single exposePreload(kind) helper.
- Three thin per-kind entries (primary.ts, settings.ts, hud.ts) each
  call exposePreload with their compile-time-known kind.
- Legacy preload.ts removed.
- New apps/desktop/src/shared/window-globals.ts centralizes the global
  names ("lightfastBridge", "codexWindowType") as typed `as const`
  exports. Renderer (main.ts, entry.tsx) reads through the constants;
  TS Window-interface augmentation keeps literal property names per TS
  requirements.

Build wiring:
- vite.preload.config.ts collapsed to a shared 13-line form (no `lib`
  block). plugin-vite's preload framework config sets
  `build.rollupOptions.input` from each forge build[].entry and emits
  via output.entryFileNames: '[name].js', so every forge entry produces
  a bundle named after its file basename — confirmed by spike.
- forge.config.ts now declares three preload build[] entries pointing at
  the shared config, replacing the single src/preload/preload.ts entry.
- factory.ts resolves the right preload bundle per WindowKind via a
  preloadFileFor() switch. The argv-based --window-kind=<kind> injection
  is gone; each preload encodes its kind at compile time.

Verification (worktree, CDP-driven via a preview of the W5 harness):
- pnpm typecheck: 52/52 packages
- pnpm check: 1104 files, no issues
- pnpm --filter @lightfast/desktop test: 45/45
- pnpm --filter @lightfast/desktop package emits bootstrap.js + primary.js
  + settings.js + hud.js (no preload.js); each preload bundle ends with
  its own kind-tagged exposePreload invocation; sizes 2885-2895 bytes.
- Live three-window CDP smoke: each window reports 15-key bridge,
  matching alphabetical key set, correct codexWindowType, correct
  dataset.windowKind, correct kind-branched React mount.

Refs:
- Plan: thoughts/shared/plans/2026-05-07-desktop-w4-preload-isolation.md
- Research: thoughts/shared/research/2026-05-06-desktop-deferred-grouping-for-worktrees.md
