+++ apps/desktop/src/preload/hud.ts
@@ -1,0 +1,3 @@
+import { exposePreload } from "./build-bridge";
+
+exposePreload("hud");
