+++ apps/desktop/src/preload/primary.ts
@@ -1,0 +1,3 @@
+import { exposePreload } from "./build-bridge";
+
+exposePreload("primary");
