+++ apps/desktop/src/preload/settings.ts
@@ -1,0 +1,3 @@
+import { exposePreload } from "./build-bridge";
+
+exposePreload("settings");
